fix: change movieSchema to verify price is integer and zodError is va…

+++ src/schemas/movies.schemas.ts
@@ -5,7 +5,7 @@ const movieSchema = z.object({
   name: z.string().min(3).max(50),
   description: z.string().nullish(),
   duration: z.number().positive(),
-  price: z.number().positive()
+  price: z.number().positive().int()
 })
 
 const movieCreateSchema = movieSchema.omit({
